Server/requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.2
flask-cors==4.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
Werkzeug==3.0.1
Flask-SQLAlchemy==3.1.1
pymysql==1.1.0
cryptography==42.0.5
python-dotenv==1.0.1
PyJWT==2.8.0
psycopg2==2.9.9